Added fn conftest and refactored as per review
