+++ tests/conftest.py
@@ -1,8 +1,9 @@
 import pytest
 
-from app.factory import create_app
-from flask import url_for
 from responses import RequestsMock
+
+from app.factory import create_app
+
 from tests.oidc_testbed import MockOIDCProvider
 
 
@@ -48,32 +49,3 @@ def app(provider):
     yield app
 
     ctx.pop()
-
-
-@pytest.fixture
-def set_email_known(browser):
-    def do_set_email_known(known=True):
-        value = 'yes' if known else 'no'
-        browser.choose('email_known', value)
-    return do_set_email_known
-
-
-@pytest.fixture
-def submit_email_address(browser):
-    def do_submit_email_address(email_address):
-        browser.fill('email_address', email_address)
-        browser.find_by_css('form button').click()
-    return do_submit_email_address
-
-
-@pytest.fixture
-def submit_known_email_address(
-        browser,
-        set_email_known,
-        submit_email_address):
-
-    def do_submit_known_email_address(email_address):
-        browser.visit(url_for('main.request_email_address', _external=True))
-        set_email_known(True)
-        submit_email_address(email_address)
-    return do_submit_known_email_address
